게시판 state lift
